+++ sdk/identity/azure_identity/src/app_service_managed_identity_credential.rs
@@ -48,6 +48,7 @@ impl AppServiceManagedIdentityCredential {
                 SECRET_ENV,
                 id,
                 client_options,
+                None,
                 env,
             ),
         }))