+++ sdk/identity/azure_identity/src/virtual_machine_managed_identity_credential.rs
@@ -3,10 +3,13 @@
 
 use crate::env::Env;
 use crate::{ImdsId, ImdsManagedIdentityCredential};
-use azure_core::http::ClientOptions;
 use azure_core::{
     credentials::{AccessToken, TokenCredential, TokenRequestOptions},
-    http::{headers::HeaderName, Url},
+    http::{
+        headers::HeaderName, ClientOptions, ExponentialRetryOptions, PipelineOptions, RetryOptions,
+        StatusCode, Url,
+    },
+    time::Duration,
 };
 use std::sync::Arc;
 
@@ -23,10 +26,38 @@ pub struct VirtualMachineManagedIdentityCredential {
 impl VirtualMachineManagedIdentityCredential {
     pub fn new(
         id: ImdsId,
-        client_options: ClientOptions,
+        mut client_options: ClientOptions,
         env: Env,
     ) -> azure_core::Result<Arc<Self>> {
         let endpoint = Url::parse(ENDPOINT).unwrap(); // valid url constant
+        let pipeline_options = Some(PipelineOptions {
+            // https://learn.microsoft.com/entra/identity/managed-identities-azure-resources/how-to-use-vm-token#error-handling
+            retry_status_codes: Vec::from([
+                StatusCode::NotFound,
+                StatusCode::Gone,
+                StatusCode::TooManyRequests,
+                StatusCode::InternalServerError,
+                StatusCode::NotImplemented,
+                StatusCode::BadGateway,
+                StatusCode::ServiceUnavailable,
+                StatusCode::GatewayTimeout,
+                StatusCode::HttpVersionNotSupported,
+                StatusCode::VariantAlsoNegotiates,
+                StatusCode::InsufficientStorage,
+                StatusCode::LoopDetected,
+                StatusCode::NotExtended,
+                StatusCode::NetworkAuthenticationRequired,
+            ]),
+            ..Default::default()
+        });
+        // these settings approximate the recommendations at
+        // https://learn.microsoft.com/entra/identity/managed-identities-azure-resources/how-to-use-vm-token#retry-guidance
+        client_options.retry = RetryOptions::exponential(ExponentialRetryOptions {
+            initial_delay: Duration::milliseconds(1340),
+            max_retries: 6,
+            max_total_elapsed: Duration::seconds(72),
+            ..Default::default()
+        });
         Ok(Arc::new(Self {
             credential: ImdsManagedIdentityCredential::new(
                 endpoint,
@@ -35,6 +66,7 @@ impl VirtualMachineManagedIdentityCredential {
                 SECRET_ENV,
                 id,
                 client_options,
+                pipeline_options,
                 env,
             ),
         }))